+++ _networkMocks/report/mockVariants/mock-report/variant20.ts
@@ -4,7 +4,7 @@ export const variant20: components['schemas']['VariantDefinition'] = {
   id: 'variantId-20',
   name: 'By Person ID and Visit Detail',
   description:
-    'Visitor and Prisoner details for those Visitors who have visits within a date range specified, along with the number of Prisoners visited and the Visitor details. This identification will facilitate multi-agency work as well as assist Establishments with their Data Cleansing.</br>VIS0019 - v0.1 Last Modified: 08 /09 / 2017',
+    'Visitor and Prisoner details for those Visitors who have visits within a date range specified, along with the number of Prisoners visited and the Visitor details. This identification will facilitate multi-agency work as well as assist Establishments with their Data Cleansing.<br>VIS0019 - v0.1 Last Modified: 08 /09 / 2017',
   resourceName: 'reports/list',
   classification: 'OFFICIAL',
   printable: false,