+++ src/dpr/components/_charts/chart/ChartTimeseries.ts
@@ -1,3 +1,4 @@
+import logger from 'src/dpr/utils/logger'
 import { components } from '../../../types/api'
 import { DashboardDataResponse } from '../../../types/Metrics'
 import {
@@ -14,6 +15,8 @@ import { PartialDate } from '../../_filters/types'
 import ChartLabelsHelper from './ChartLabels'
 
 class TimeseriesChart {
+  id!: string
+
   labels: string[] = []
 
   datasets: DashboardVisualisationDataSet[] = []
@@ -55,6 +58,9 @@ class TimeseriesChart {
   }
 
   initFromDefinition = (definition: BarTimeseriesDefinitionType | LineTimeseriesDefinitionType) => {
+    logger.debug('Building Dashboard Visualisation: ', JSON.stringify(definition))
+
+    this.id = definition.id
     this.measures = definition.columns.measures
     this.keys = definition.columns.keys || []
     this.unit = this.measures.find(m => m.unit)?.unit
@@ -75,7 +81,8 @@ class TimeseriesChart {
     )
 
     if (!dateMeasure) {
-      throw new Error('No date field in definition')
+      logger.debug(this.id, JSON.stringify(this.measures))
+      throw new Error(`${this.id}: No timestamp field in definition`)
     }
 
     this.timeBlockData = DatasetHelper.groupRowsByTimestamp(this.responseData, dateMeasure)
@@ -85,12 +92,20 @@ class TimeseriesChart {
 
   buildDatasets = () => {
     this.chartColoursHelper = new ChartColoursHelper()
+
     for (let index = 0; index < this.datasetCount; index += 1) {
       const data = this.timeBlockData.map(timeperiod => {
-        const valueId = this.measures[1].id
-        const period = timeperiod[index]
-        return period && period[valueId].raw ? Number(period[valueId].raw) : 0
+        const valueId = this.measures[1]?.id
+        const value = timeperiod[index]?.[valueId]?.raw
+
+        if (value == null || valueId == null) {
+          return 0
+        }
+
+        const numericValue = Number(value)
+        return Number.isNaN(numericValue) ? 0 : numericValue
       })
+
       const total = data.reduce((a, c) => a + c, 0)
       const label = this.chartLabelsHelper.getDatasetLabel(this.keys, this.timeBlockData[0][index])