+++ src/dpr/components/_charts/utils.ts
@@ -4,16 +4,16 @@ import { components } from '../../types/api'
 import { ChartDetails, ChartMetaData } from '../../types/Charts'
 import { DashboardDataResponse } from '../../types/Metrics'
 import DatasetHelper, {
+  getDateValue,
   getTimestampColumn,
   getTimestampMeasure,
-  getDateValue,
 } from '../../utils/Dashboards/VisualisationDatasetHelper'
 import { mapUnitToSymbol } from '../../utils/Dashboards/VisualisationUnitHelper'
 import DashboardListUtils from '../_dashboards/dashboard-list/utils'
 import {
+  DashboardVisualisationCardData,
   DashboardVisualisationData,
   DashboardVisualisationType,
-  DashboardVisualisatonCardData,
   MoJTable,
 } from '../_dashboards/dashboard-visualisation/types'
 import { UnitType } from '../_dashboards/dashboard-visualisation/Validate'
@@ -32,7 +32,7 @@ export const createChart = (
   chartDefinition: components['schemas']['DashboardVisualisationDefinition'],
   rawData: DashboardDataResponse[],
   type: components['schemas']['DashboardVisualisationDefinition']['type'],
-): DashboardVisualisatonCardData | undefined => {
+): DashboardVisualisationCardData | undefined => {
   let table: MoJTable | undefined
   let chart: DashboardVisualisationData | undefined
   let details: ChartDetails | undefined
@@ -202,7 +202,7 @@ const createHeadlines = (
 
     if (headlineColumn) {
       const { id } = headlineColumn
-      const { raw } = data[0][id]
+      const rawValue = data[0][id]?.raw
 
       const dateColumn = getTimestampMeasure(measures)
       const dateData = getDateValue(data, dateColumn)
@@ -211,7 +211,8 @@ const createHeadlines = (
         label = dateData.value
       }
 
-      value = raw ? Number(raw) : undefined
+      const numericValue = Number(rawValue)
+      value = Number.isNaN(numericValue) ? undefined : numericValue
 
       if (value) {
         headline = {