+++ src/dpr/components/_dashboards/dashboard-visualisation/Validate.ts
@@ -71,7 +71,7 @@ export const DashboardVisualisationSchema = z.object({
     'matrix-timeseries',
     'line-timeseries',
   ]),
-  display: z.string().optional(),
+  display: z.string(),
   description: z.string().optional(),
   options: z.object(dashboardOptions.shape).optional(),
   columns: z.object(DashboardColumns.shape),