+++ src/dpr/components/truncate/utils.ts
@@ -1,0 +1,49 @@
+import truncate from 'truncate-html'
+import nunjucks from 'nunjucks'
+import { sanitiseHtml } from 'src/dpr/utils/sanitizeHtml'
+
+export interface TruncationOptions {
+  stringValue: string
+  charLength?: number
+  showMore?: boolean
+  classes?: string
+}
+
+export interface TruncationModel {
+  fullHtml: string
+  truncatedHtml: string
+  showMore: boolean
+  isTruncated: boolean
+  classes?: string | undefined
+}
+
+export const initialiseTruncation = ({
+  stringValue,
+  charLength = 80,
+  showMore = true,
+  classes,
+}: TruncationOptions): TruncationModel => {
+  const sanitizedHtml = sanitiseHtml(stringValue)
+
+  const truncatedHtml = truncate(sanitizedHtml, {
+    length: charLength,
+  })
+
+  const isTruncated = truncatedHtml !== sanitizedHtml
+
+  return {
+    fullHtml: sanitizedHtml,
+    truncatedHtml,
+    showMore,
+    isTruncated,
+    classes,
+  }
+}
+
+export const renderTruncateAsHtml = (options: TruncationOptions): string => {
+  const params = initialiseTruncation(options)
+
+  return nunjucks.render('dpr/components/truncate/render.njk', {
+    params,
+  })
+}