+++ src/dpr/routes/journeys/view-report/async/report/routes.ts
@@ -1,21 +1,22 @@
 import { Router } from 'express'
-import ViewAsyncReportController from './controller'
-import ViewReportController from '../controller'
+import { ReportType } from 'src/dpr/types/UserReports'
+import { ViewAsyncReportController } from './controller'
+import { ViewAsyncController } from '../controller'
 import { Services } from '../../../../../types/Services'
 import { validateFilters } from '../../../../../validation/validateFilters'
 
 export function routes({ layoutPath, services }: { layoutPath: string; services: Services }): Router {
   const router = Router({ mergeParams: true })
   const asyncReportController = new ViewAsyncReportController(layoutPath, services)
-  const viewReportController = new ViewReportController(layoutPath, services)
+  const viewReportController = new ViewAsyncController(layoutPath, services)
 
-  router.get([`/`, `/download-disabled`], asyncReportController.GET)
+  router.get([`/`, `/download-disabled`], viewReportController.GET(ReportType.REPORT))
 
   // ----------------------------
   // FILTERS
   // ----------------------------
   router.post('/apply-filters', validateFilters({ interactive: true }), asyncReportController.applyFilters)
-  router.post('/reset-filters', asyncReportController.resetFilters)
+  router.post('/reset-filters', viewReportController.resetFilters)
 
   // ---------------------------
   // COLUMNS