+++ test-app/mocks/mockClients/reports/mockVariants/report-templates/parent-child/parent-child.js
@@ -2,7 +2,7 @@
 const reportTemplateExampleParentChild = {
   id: 'report-template-parent-child',
   name: 'Parent-child',
-  description: 'Template: "parent-child".</br> A report with parent and child datasets',
+  description: 'Template: "parent-child".<br> A report with parent and child datasets',
   resourceName: 'reports/list',
   classification: 'OFFICIAL',
   printable: true,