+++ test-app/routes/components/routes.ts
@@ -7,6 +7,7 @@ import DashboardRoutes from './dashboards/routes'
 import UserReportsRoutes from './user-reports/routes'
 import FiltersRoutes from './filters/routes'
 import { ReportHeadingRoutes } from './report-heading/routes'
+import TruncateRoutes from './truncate/routes'
 
 import ComponentsController from './controller'
 import { Services } from '../../../src/dpr/types/Services'
@@ -23,6 +24,7 @@ export function Routes(services: Services): Router {
   router.use(`/dashboards`, DashboardRoutes())
   router.use(`/filters`, FiltersRoutes())
   router.use('/report-heading', ReportHeadingRoutes())
+  router.use('/truncate', TruncateRoutes())
 
   return router
 }